requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
bcrypt==3.2.2
passlib[bcrypt]
pydantic
pydantic[email]
python-dotenv
python-multipart